initial structure changes for storing PRs
