+++ bin/config.js
@@ -27,7 +27,8 @@ function loadFactory(spec, namespace) {
 module.exports = {
   summary: {},
   curation: {
-    service: loadFactory(config.get('CURATION_PROVIDER') || 'github', 'curation')
+    service: loadFactory(config.get('CURATION_PROVIDER') || 'github', 'curation.service'),
+    store: loadFactory(config.get('CURATION_STORE_PROVIDER') || 'github', 'curation.store')
   },
   harvest: {
     service: loadFactory(config.get('HARVESTER_PROVIDER') || 'crawler', 'harvest.service'),
