+++ package.json
@@ -53,7 +53,7 @@
     "semver": "^5.4.1",
     "serialize-error": "^2.1.0",
     "spdx-license-list": "^4.1.0",
-    "swagger-ui-express": "^2.0.13",
+    "swagger-ui-express": "^4.0.1",
     "throat": "^4.1.0",
     "tmp": "0.0.33",
     "vso-node-api": "^6.2.8-preview"
