+++ providers/curation/githubConfig.js
@@ -2,17 +2,28 @@
 // SPDX-License-Identifier: MIT
 const config = require('painless-config')
 const githubService = require('./github')
+const githubStore = require('./githubStore')
 
-function github(options, endpoints) {
+const owner = config.get('CURATION_GITHUB_OWNER') || 'clearlydefined'
+const repo = config.get('CURATION_GITHUB_REPO') || 'curated-data'
+const branch = config.get('CURATION_GITHUB_BRANCH') || 'master'
+const token = config.get('CURATION_GITHUB_TOKEN')
+
+function service(options, store, endpoints) {
+  const realOptions = options || { owner, repo, branch, token }
+  return githubService(realOptions, store, endpoints)
+}
+
+function store(options) {
   const realOptions = options || {
-    owner: config.get('CURATION_GITHUB_OWNER') || 'clearlydefined',
-    repo: config.get('CURATION_GITHUB_REPO') || 'curated-data',
-    branch: config.get('CURATION_GITHUB_BRANCH') || 'master',
-    token: config.get('CURATION_GITHUB_TOKEN'),
+    owner,
+    repo,
+    branch,
+    token,
     tempLocation: config.get('CURATION_TEMP_LOCATION') || (process.platform === 'win32' ? 'c:/temp' : '/tmp'),
     curationFreshness: config.get('CURATION_FRESHNESS') || 600000
   }
-  return githubService(realOptions, endpoints)
+  return githubStore(realOptions)
 }
 
-module.exports = github
+module.exports = { service, store }
